[ refactor ] 헤더구조 변경(직관적으로 어떻게 구성되어있는지 알기쉽게 children 속성으로 넣음)
