+++ src/@components/main/index.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import styled, { css } from "styled-components";
-import { MainTracksTextIc, MainVocalsTextIc } from "../../assets";
+import { MainTracksTextIc, MainVocalsTextIc, SloganIc } from "../../assets";
 import hoverTracksBackground from "../../assets/image/hoverTracksImg.png";
 import hoverVocalsBackground from "../../assets/image/hoverVocalsImg.png";
 import originBackground from "../../assets/image/mainBackgroundImg.png";
@@ -13,6 +13,7 @@ import Footer from "../@common/footer";
 import Header from "../@common/header";
 import Ads from "./ads";
 import LoginButton from "./loginButton";
+import HomeLogo from "../@common/homeLogo";
 
 export default function Main() {
   const showModal = useRecoilValue(openConventionModal);
@@ -57,7 +58,9 @@ export default function Main() {
 
   return (
     <>
-      <Header homeLogo slogan headerStyle={headerStyle}>
+      <Header headerStyle={headerStyle}>
+        <HomeLogo />
+        <SloganIcon />
         <LoginButton />
       </Header>
       <MainPageWrapper>
@@ -83,7 +86,6 @@ export default function Main() {
 
 const headerStyle: React.CSSProperties = {
   position: "absolute",
-  
 };
 
 const Background = styled.main<{
@@ -171,6 +173,12 @@ const MainSlogan = styled.img`
 
   top: 50.6rem;
   left: 8.7rem;
+
+  width: 35rem;
+`;
+
+const SloganIcon = styled(SloganIc)`
+  cursor: pointer;
 
   width: 35rem;
 `;
