+++ src/@components/upload/index.tsx
@@ -1,3 +1,4 @@
+import BackButton from "../@common/backButton";
 import Header from "../@common/header";
 import UploadBody from "./uploadBody";
 import UploadHeader from "./uploadHeader";
@@ -5,9 +6,11 @@ import UploadHeader from "./uploadHeader";
 export default function UploadContainer() {
   return (
     <>
-      <Header backBtn prevURL="-1">
+      <Header>
+        <BackButton staticPrevURL={-1} />
         <UploadHeader />
       </Header>
+
       <UploadBody />
     </>
   );
