+++ src/@components/uploadEdit/index.tsx
@@ -1,3 +1,4 @@
+import BackButton from "../@common/backButton";
 import Header from "../@common/header";
 import UploadBody from "../upload/uploadBody";
 import UploadHeader from "../upload/uploadHeader";
@@ -5,7 +6,8 @@ import UploadHeader from "../upload/uploadHeader";
 export default function UploadEditContainer() {
   return (
     <>
-      <Header backBtn>
+      <Header>
+        <BackButton />
         <UploadHeader />
       </Header>
       <UploadBody />
