+++ src/@pages/loginPage.tsx
@@ -1,10 +1,14 @@
+import BackButton from "../@components/@common/backButton";
 import Header from "../@components/@common/header";
 import LoginForm from "../@components/login/loginForm";
 
 export default function LoginPage() {
   return (
     <>
-      <Header backBtn prevURL="/" />
+      <Header>
+        <BackButton />
+      </Header>
+
       <LoginForm />
     </>
   );
