+++ src/@pages/vocalSearchPage.tsx
@@ -6,6 +6,7 @@ import { PlayerProvider } from "../context/playerContext";
 import TrackSearchHeader from "../@components/trackSearch/trackSearchHeader/trackSearchHeader";
 
 import Header from "../@components/@common/header";
+import HomeLogo from "../@components/@common/homeLogo";
 
 const Wrapper = styled.section`
   display: flex;
@@ -14,7 +15,8 @@ export default function VocalSearchPage() {
   return (
     <>
       <PlayerProvider>
-        <Header homeLogo>
+        <Header>
+          <HomeLogo />
           <TrackSearchHeader pageType="vocals" />
         </Header>
         <Wrapper>
